now there are substates of admin to manage vaious tasks, all order templates and functions moved to separate child component, now we will start building content component
