+++ src/components/admin/login/login.html
@@ -1,0 +1,1 @@
+<h1>Please login</h1>
