+++ src/components/admin/login/login.js
@@ -1,0 +1,5 @@
+import template from './login.html';
+
+export default({
+    template,
+});
